requirements.txt
agno
python-dotenv
OPENAI
Groq
duckduckgo-search
yfinance
ddgs
fastapi
uvicorn